Create 05-Find The Missing Number In A Sorted Array
Given a Sorted Array of n-1 integers and these integers are in the range of 1 to n. There are no duplicates in list. One of the integers is missing in the list. Write an efficient code to find the missing integer. 

Examples: 

Input : arr[] = [1, 2, 3, 4, 6, 7, 8]
Output : 5

Input : arr[] = [1, 2, 3, 4, 5, 6, 8, 9]
Output : 7

Input : arr[] = [2, 3, 4, 5, 6, 7, 8]
Output : 1

Input : arr[] = [1, 2, 3, 4, 5, 6, 7, 9]
Output : 8