Update 29-Find Missing In Second Array
Given two arrays A and B contains integers of size N and M, the task is to find numbers which are present in the first array, but not present in the second array.

(https://practice.geeksforgeeks.org/problems/in-first-but-second5423/1?utm_source=gfg&utm_medium=article&utm_campaign=bottom_sticky_on_article)

Example 1:

Input: M = 6, N = 5
A[] = {1, 2, 3, 4, 5, 10}
B[] = {2, 3, 1, 0, 5}
Output: 4 10
Explanation: 4 and 10 are present in 
first array, but not in second array.
Example 2:

Input: M = 5, N = 5
A[] = {4, 3, 5, 9, 11}
B[] = {4, 9, 3, 11, 10}
Output: 5  
Explanation: Second array does not 
contain element 5.


Input: M = 24, N = 6
A[] = {227, 2, 17, 15, 17, 21, 26, 30, 16, 12, 19, 22, 18, 8, 24, 10, 17, 28, 14, 7, 4, 21, 3, 4}

B[] = {25, 12, 30, 24, 14, 34}

Output:  227, 2, 17, 15, 17, 21, 26, 16, 19, 22, 18, 8, 10, 17, 28, 7, 4, 21, 3, 4 

Explanation: Second array does not 
contain the aboev elements .

Your Task:
This is a function problem. You don't need to take any input, as it is already accomplished by the driver code. You just need to complete the function findMissing() that takes array A, array B, integer N, and integer M  as parameters and returns an array that contains the missing elements and the order of the elements should be the same as they are in array A.

 

Expected Time Complexity: O(N+M).
Expected Auxiliary Space: O(M).
 

Constraints:
1 ≤ N, M ≤ 10^6
-10^6 ≤ A[i], B[i] ≤ 10^6